+++ AssemblyTasks.vb
@@ -1687,6 +1687,8 @@ Public Class AssemblyTasks
 
         Dim OutString As String = ""
 
+        Dim FCD As New FilenameCharmapDoctor()
+
         ' Formatting for subdirectory name formula
         ' Example property callout: %{hmk_Part_Number/CP|G}  
         ' Need to know PropertySet, so maybe: %{Custom.hmk_Part_Number}
@@ -1799,6 +1801,7 @@ Public Class AssemblyTasks
         End If
 
         If ExitStatus = 0 Then
+            OutString = FCD.SubstituteIllegalCharacters(OutString)
             ErrorMessageList.Add(OutString)
         End If