+++ PartTasks.vb
@@ -1663,6 +1663,8 @@ Public Class PartTasks
 
         Dim OutString As String = ""
 
+        Dim FCD As New FilenameCharmapDoctor()
+
         ' Formatting for subdirectory name formula
         ' Example property callout: %{hmk_Part_Number/CP|G}  
         ' Need to know PropertySet, so maybe: %{Custom.hmk_Part_Number}
@@ -1775,6 +1777,7 @@ Public Class PartTasks
         End If
 
         If ExitStatus = 0 Then
+            OutString = FCD.SubstituteIllegalCharacters(OutString)
             ErrorMessageList.Add(OutString)
         End If